@@ -32,8 +32,8 @@ endif()
 # Tell cmake to export a compile_commands.json file
 set(CMAKE_EXPORT_COMPILE_COMMANDS 1)
 
-# Tell cmake we need C++17
-set(CMAKE_CXX_STANDARD 17)
+# Tell cmake we need C++20
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 ## Setup compiler flags for various build types
