@@ -4,13 +4,15 @@
 #include "assets/AssetManager.hh"
 #include "core/Logging.hh"
 
+#include <filesystem>
 #include <fstream>
 #include <glm/gtc/type_ptr.hpp>
 #include <iostream>
 #include <murmurhash/MurmurHash3.h>
+#include <tiny_gltf.h>
 
 namespace sp {
-    glm::mat4 GetNodeMatrix(tinygltf::Node *node) {
+    glm::mat4 GetNodeMatrix(const tinygltf::Node *node) {
         glm::mat4 out(1.0);
 
         if (node->matrix.size() == 16) {
@@ -33,11 +35,11 @@ namespace sp {
         return out;
     }
 
-    Model::Attribute GetPrimitiveAttribute(shared_ptr<tinygltf::Model> model,
-                                           tinygltf::Primitive *p,
-                                           string attribute) {
+    Model::Attribute GetPrimitiveAttribute(std::shared_ptr<const tinygltf::Model> model,
+                                           const tinygltf::Primitive *p,
+                                           std::string attribute) {
         if (!p->attributes.count(attribute)) return Model::Attribute();
-        auto accessor = model->accessors[p->attributes[attribute]];
+        auto accessor = model->accessors[p->attributes.at(attribute)];
         auto bufView = model->bufferViews[accessor.bufferView];
 
         size_t componentCount = 1;
@@ -61,33 +63,111 @@ namespace sp {
                                 bufView.buffer};
     }
 
-    Model::Model(const std::string &name, shared_ptr<tinygltf::Model> model) : name(name), model(model) {
-        int defaultScene = 0;
-
-        if (model->defaultScene != -1) { defaultScene = model->defaultScene; }
-
-        for (int node : model->scenes[defaultScene].nodes) {
-            AddNode(node, glm::mat4());
-        }
-    }
-
     Model::~Model() {
         Debugf("Destroying model %s", name);
     }
 
     bool Model::HasBuffer(size_t index) const {
+        Assert(valid.test(), "Checking buffer on invalid model");
         return model->buffers.size() > index;
     }
 
-    const vector<unsigned char> &Model::GetBuffer(size_t index) const {
+    const std::vector<unsigned char> &Model::GetBuffer(size_t index) const {
+        Assert(valid.test(), "Accessing buffer on invalid model");
         return model->buffers[index].data;
     }
 
     Hash128 Model::HashBuffer(size_t index) const {
+        Assert(valid.test(), "Hashing buffer on invalid model");
         Hash128 output;
         auto buffer = GetBuffer(index);
         MurmurHash3_x86_128(buffer.data(), buffer.size(), 0, output.data());
         return output;
+    }
+
+    // Returns a vector of the GLTF node indexes that are present in the "joints"
+    // array of the GLTF skin.
+    std::vector<int> Model::GetJointNodes() const {
+        Assert(valid.test(), "Getting joint nodes on invalid model");
+        std::vector<int> nodes;
+
+        // TODO: deal with GLTFs that have more than one skin
+        for (int node : model->skins[0].joints) {
+            nodes.push_back(node);
+        }
+
+        return nodes;
+    }
+
+    int Model::FindNodeByName(std::string name) const {
+        Assert(valid.test(), "Finding joint nodes on invalid model");
+        for (size_t i = 0; i < model->nodes.size(); i++) {
+            if (model->nodes[i].name == name) { return i; }
+        }
+        return -1;
+    }
+
+    glm::mat4 Model::GetInvBindPoseForNode(int nodeIndex) const {
+        Assert(valid.test(), "Getting bind pose on invalid model");
+        return inverseBindMatrixForJoint.at(nodeIndex);
+    }
+
+    std::string Model::GetNodeName(int node) const {
+        Assert(valid.test(), "Getting node name on invalid model");
+        return model->nodes[node].name;
+    }
+
+    void Model::PopulateFromAsset(std::shared_ptr<const Asset> asset, const tinygltf::FsCallbacks *fsCallbacks) {
+        Assert(asset != nullptr, "Loading Model from null asset");
+        this->asset = asset;
+        asset->WaitUntilValid();
+
+        std::filesystem::path fsPath(asset->path);
+        std::string baseDir = fsPath.remove_filename().string();
+
+        Debugf("Loading model %s from %s", name, asset->path);
+
+        auto gltfModel = std::make_shared<tinygltf::Model>();
+        std::string err;
+        std::string warn;
+        bool ret = false;
+        tinygltf::TinyGLTF gltfLoader;
+        if (fsCallbacks != nullptr) {
+            gltfLoader.SetFsCallbacks(*fsCallbacks);
+        } else {
+            tinygltf::FsCallbacks fs = {};
+            gltfLoader.SetFsCallbacks(fs);
+        }
+        if (ends_with(asset->path, ".gltf")) {
+            ret = gltfLoader.LoadASCIIFromString(gltfModel.get(),
+                                                 &err,
+                                                 &warn,
+                                                 (char *)asset->Buffer(),
+                                                 asset->BufferSize(),
+                                                 baseDir);
+        } else {
+            // Assume GLB model
+            ret = gltfLoader.LoadBinaryFromMemory(gltfModel.get(),
+                                                  &err,
+                                                  &warn,
+                                                  (unsigned char *)asset->Buffer(),
+                                                  asset->BufferSize(),
+                                                  baseDir);
+        }
+
+        Assert(err.empty(), "Failed to parse glTF: " + err);
+        Assert(ret, "Failed to parse glTF");
+        this->model = gltfModel;
+
+        int defaultScene = 0;
+        if (model->defaultScene != -1) { defaultScene = model->defaultScene; }
+
+        for (int node : model->scenes[defaultScene].nodes) {
+            AddNode(node, glm::mat4());
+        }
+
+        this->valid.test_and_set();
+        this->valid.notify_all();
     }
 
     void Model::AddNode(int nodeIndex, glm::mat4 parentMatrix) {
@@ -186,33 +266,5 @@ namespace sp {
         for (int childNodeIndex : node.children) {
             AddNode(childNodeIndex, matrix);
         }
-    }
-
-    // Returns a vector of the GLTF node indexes that are present in the "joints"
-    // array of the GLTF skin.
-    std::vector<int> Model::GetJointNodes() const {
-        std::vector<int> nodes;
-
-        // TODO: deal with GLTFs that have more than one skin
-        for (int node : model->skins[0].joints) {
-            nodes.push_back(node);
-        }
-
-        return nodes;
-    }
-
-    int Model::FindNodeByName(std::string name) const {
-        for (size_t i = 0; i < model->nodes.size(); i++) {
-            if (model->nodes[i].name == name) { return i; }
-        }
-        return -1;
-    }
-
-    glm::mat4 Model::GetInvBindPoseForNode(int nodeIndex) const {
-        return inverseBindMatrixForJoint.at(nodeIndex);
-    }
-
-    std::string Model::GetNodeName(int node) const {
-        return model->nodes[node].name;
     }
 } // namespace sp
